Classをファイルごとに分割，コメントのインデントを修正
